Refactor: css naming
